-Using unsigned short instead of uint8_t for buidlings coords -Working on citymaker, a parser for creating cities with text files  (this is useful for debugging and for creating actual cities) -Minor fixes and improvements

+++ structs.h
@@ -22,7 +22,7 @@ typedef struct geo_pos{
 
 
 struct building {
-	uint8_t x,y;
+	unsigned short x,y;
 	unsigned short side_size;
 	unsigned short type;
 	void* item;
